requirements.txt
djangorestframework==3.6.3
pkg-resources==0.0.0
